abstract examples from demo.c for #3
examples and their utils have been moved to examples.h
