+++ src/demo/examples.h
@@ -1,0 +1,14 @@
+#ifndef _EXAMPLES_H_
+#define _EXAMPLES_H_
+
+#include <stddef.h>
+
+// Examples definition
+
+size_t japan_example(void);
+size_t line_example(void);
+size_t triangle_example(void);
+size_t blend_example(void);
+size_t text_example(void);
+
+#endif // _EXAMPLES_H_
